backend/requirements.txt
asgiref==3.9.1
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
curl_cffi==0.13.0
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
dotenv==0.9.9
drf-spectacular==0.28.0
frozendict==2.4.6
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
multitasking==0.0.12
numpy==2.2.6
openapi-codec==1.3.2
packaging==25.0
pandas==2.3.2
peewee==3.18.2
platformdirs==4.5.0
protobuf==6.32.1
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
simplejson==3.20.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.66
celery==5.4.0
redis==5.2.1
scipy==1.15.3
setuptools>=70.0,<71.0
